+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.12 ] - 2023/02/01
+
+  * Updated `SpotifyClient._CheckResponseForErrors` method with better logging information.
 
 ###### [ 1.0.11 ] - 2023/02/01